+++ goldens/circular-deps/packages.json
@@ -177,73 +177,12 @@
     "packages/core/src/change_detection/differs/keyvalue_differs.ts"
   ],
   [
-    "packages/core/src/di.ts",
-    "packages/core/src/di/index.ts",
-    "packages/core/src/di/injectable.ts",
-    "packages/core/src/di/jit/injectable.ts",
-    "packages/core/src/di/jit/environment.ts",
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts",
-    "packages/core/src/di/metadata.ts",
-    "packages/core/src/render3/instructions/di.ts"
-  ],
-  [
-    "packages/core/src/di.ts",
-    "packages/core/src/di/index.ts",
-    "packages/core/src/di/injectable.ts",
-    "packages/core/src/di/jit/injectable.ts",
-    "packages/core/src/di/jit/environment.ts",
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/metadata.ts",
-    "packages/core/src/render3/instructions/di.ts"
-  ],
-  [
-    "packages/core/src/di.ts",
-    "packages/core/src/di/index.ts",
-    "packages/core/src/di/injectable.ts",
-    "packages/core/src/di/jit/injectable.ts",
-    "packages/core/src/di/jit/util.ts",
-    "packages/core/src/di/metadata.ts",
-    "packages/core/src/render3/instructions/di.ts"
-  ],
-  [
-    "packages/core/src/di.ts",
-    "packages/core/src/di/index.ts",
-    "packages/core/src/di/metadata.ts",
-    "packages/core/src/render3/instructions/di.ts"
-  ],
-  [
-    "packages/core/src/di.ts",
-    "packages/core/src/di/index.ts",
-    "packages/core/src/di/reflective_injector.ts",
-    "packages/core/src/di/metadata.ts",
-    "packages/core/src/render3/instructions/di.ts"
-  ],
-  [
-    "packages/core/src/di.ts",
-    "packages/core/src/di/index.ts",
-    "packages/core/src/di/reflective_injector.ts",
-    "packages/core/src/di/reflective_provider.ts",
-    "packages/core/src/di/metadata.ts",
-    "packages/core/src/render3/instructions/di.ts"
-  ],
-  [
     "packages/core/src/di/injectable.ts",
     "packages/core/src/di/jit/injectable.ts"
   ],
   [
     "packages/core/src/di/injector_compatibility.ts",
     "packages/core/src/di/injector.ts"
-  ],
-  [
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts",
-    "packages/core/src/di/null_injector.ts"
-  ],
-  [
-    "packages/core/src/di/injector_compatibility.ts",
-    "packages/core/src/di/injector.ts",
-    "packages/core/src/di/r3_injector.ts"
   ],
   [
     "packages/core/src/di/injector_token.ts",
